code modified on access control the user can not edit or delet others post

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,16 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -86,6 +96,12 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+
+        //Check for correct user
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -122,6 +138,13 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+
+        //Check for correct user
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized page');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success','Post Deleted');
